+++ sugarloaf/src/font/fonts.rs
@@ -8,13 +8,30 @@ pub enum SugarloafFontStyle {
     Italic,
 }
 
+#[derive(Debug, Default, Serialize, Deserialize, PartialEq, Clone)]
+pub enum SugarloafFontWidth {
+    UltraCondensed,
+    ExtraCondensed,
+    Condensed,
+    SemiCondensed,
+    #[default]
+    Normal,
+    SemiExpanded,
+    Expanded,
+    ExtraExpanded,
+    UltraExpanded,
+}
+
 #[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
 pub struct SugarloafFont {
     #[serde(default = "default_font_family")]
     pub family: String,
+    #[serde(default = "Option::default")]
     pub weight: Option<u16>,
     #[serde(default = "SugarloafFontStyle::default")]
     pub style: SugarloafFontStyle,
+    #[serde(default = "Option::default")]
+    pub width: Option<SugarloafFontWidth>,
 }
 
 impl Default for SugarloafFont {
@@ -23,6 +40,7 @@ impl Default for SugarloafFont {
             family: default_font_family(),
             weight: None,
             style: SugarloafFontStyle::Normal,
+            width: None,
         }
     }
 }
@@ -52,8 +70,9 @@ fn default_font_family() -> String {
 pub fn default_font_regular() -> SugarloafFont {
     SugarloafFont {
         family: default_font_family(),
-        weight: Some(300),
+        weight: Some(400),
         style: SugarloafFontStyle::Normal,
+        width: None,
     }
 }
 
@@ -62,6 +81,7 @@ pub fn default_font_bold() -> SugarloafFont {
         family: default_font_family(),
         weight: Some(800),
         style: SugarloafFontStyle::Normal,
+        width: None,
     }
 }
 
@@ -70,6 +90,7 @@ pub fn default_font_italic() -> SugarloafFont {
         family: default_font_family(),
         weight: Some(300),
         style: SugarloafFontStyle::Italic,
+        width: None,
     }
 }
 
@@ -78,6 +99,7 @@ pub fn default_font_bold_italic() -> SugarloafFont {
         family: default_font_family(),
         weight: Some(800),
         style: SugarloafFontStyle::Italic,
+        width: None,
     }
 }