+++ sugarloaf/src/font/mod.rs
@@ -7,7 +7,7 @@ pub mod loader;
 pub const FONT_ID_REGULAR: usize = 0;
 
 use crate::font::constants::*;
-use crate::font::fonts::SugarloafFontStyle;
+use crate::font::fonts::{SugarloafFontStyle, SugarloafFontWidth};
 use crate::font_introspector::text::cluster::Parser;
 use crate::font_introspector::text::cluster::Token;
 use crate::font_introspector::text::cluster::{CharCluster, Status};
@@ -356,6 +356,7 @@ impl FontLibraryData {
                     family: extra_font.family,
                     style: extra_font.style,
                     weight: extra_font.weight,
+                    width: extra_font.width,
                 },
                 true,
                 true,
@@ -565,6 +566,32 @@ fn find_font(
 
         if let Some(weight) = font_spec.weight {
             query.weight = crate::font::loader::Weight(weight);
+        }
+
+        if let Some(ref width) = font_spec.width {
+            query.stretch = match width {
+                SugarloafFontWidth::UltraCondensed => {
+                    crate::font::loader::Stretch::UltraCondensed
+                }
+                SugarloafFontWidth::ExtraCondensed => {
+                    crate::font::loader::Stretch::ExtraCondensed
+                }
+                SugarloafFontWidth::Condensed => crate::font::loader::Stretch::Condensed,
+                SugarloafFontWidth::SemiCondensed => {
+                    crate::font::loader::Stretch::SemiCondensed
+                }
+                SugarloafFontWidth::Normal => crate::font::loader::Stretch::Normal,
+                SugarloafFontWidth::SemiExpanded => {
+                    crate::font::loader::Stretch::SemiExpanded
+                }
+                SugarloafFontWidth::Expanded => crate::font::loader::Stretch::Expanded,
+                SugarloafFontWidth::ExtraExpanded => {
+                    crate::font::loader::Stretch::ExtraExpanded
+                }
+                SugarloafFontWidth::UltraExpanded => {
+                    crate::font::loader::Stretch::UltraExpanded
+                }
+            };
         }
 
         query.style = match font_spec.style {